Fixed running tasks from tests
